extract Group Ajax
